requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.56.0
accelerate>=1.10.0
peft>=0.17.0
datasets>=4.0.0
timm>=1.0.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
reportlab>=4.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
safetensors>=0.6.0
